+++ sonic-xcvrd/setup.py
@@ -1,35 +1,4 @@
 from setuptools import setup, find_packages
-from distutils.command.build_py import build_py as _build_py
-import distutils.command
-import pkg_resources
-#import subprocess
-#proto_files = [os.path.join(os.path.dirname(os.path.abspath(__file__)),'proto', 'image_transform.proto')]
-#proto_files = ['proto/image_transform.proto']
-
-class GrpcTool(distutils.cmd.Command):
-    def initialize_options(self):
-        pass
-
-    def finalize_options(self):
-        pass
-
-    def run(self):
-        import grpc_tools.protoc
-
-        proto_include = pkg_resources.resource_filename('proto', '')
-
-        grpc_tools.protoc.main([
-            'grpc_tools.protoc',
-            '-I{}'.format(proto_include),
-            '--python_out=proto_out/',
-            '--grpc_python_out=proto_out/',
-            'linkmgr_grpc_driver.proto'
-        ])
-
-class BuildPyCommand (_build_py, object):
-    def run(self):
-        self.run_command('GrpcTool')
-        super(BuildPyCommand, self).run()
 
 setup(
     name='sonic-xcvrd',
@@ -47,16 +16,13 @@ setup(
             'xcvrd = xcvrd.xcvrd:main',
         ]
     },
-    cmdclass={'build_py': BuildPyCommand,
-              'GrpcTool': GrpcTool},
     install_requires=[
         # NOTE: This package also requires swsscommon, but it is not currently installed as a wheel
         'enum34; python_version < "3.4"',
         'sonic-py-common',
     ],
     setup_requires=[
-        'wheel',
-        'grpcio-tools'
+        'wheel'
     ],
     tests_require=[
         'pytest',
